requirements.txt
flask
torch
torchvision
timm
pillow
google-generativeai
